Spoofed ip and user agent. Refactored class code into their own files
