+++ ds_scraper/get_fake_user_proxies.py
@@ -11,51 +11,30 @@ import random
 from get_fake_user_agent import getRandomUserAgent
 
 ua = getRandomUserAgent()
+#ua = 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/67.0.3396 Safari/537.36'
 proxies = [] # Will contain proxies [ip, port]
 
-# Main function
-def main():
-  # Retrieve latest proxies
-  proxies_req = Request('https://www.sslproxies.org/')
-  proxies_req.add_header('User-Agent', ua)
-  proxies_doc = urlopen(proxies_req).read().decode('utf8')
 
-  soup = BeautifulSoup(proxies_doc, 'html.parser')
-  proxies_table = soup.find(id='proxylisttable')
+def getProxy():
+    proxies_req = Request('https://www.sslproxies.org/')
+    proxies_req.add_header('User-Agent', ua)
+    proxies_doc = urlopen(proxies_req).read().decode('utf8')
 
-  # Save proxies in the array
-  for row in proxies_table.tbody.find_all('tr'):
-    proxies.append({
-      'ip':   row.find_all('td')[0].string,
-      'port': row.find_all('td')[1].string
-    })
+    soup = BeautifulSoup(proxies_doc, 'html.parser')
+    proxies_table = soup.find(id='proxylisttable')
 
-  # Choose a random proxy
-  proxy_index = random_proxy()
-  proxy = proxies[proxy_index]
+    # Save proxies in the array
+    for row in proxies_table.tbody.find_all('tr'):
+        proxies.append({
+                'ip':   row.find_all('td')[0].string,
+                'port': row.find_all('td')[1].string
+                })
 
-  for n in range(1, 100):
-    req = Request('http://icanhazip.com')
-    req.set_proxy(proxy['ip'] + ':' + proxy['port'], 'http')
+    # Choose a random proxy
+    proxy_index = random_proxy()
+    proxy = proxies[proxy_index]
+    
+    return proxy
 
-    # Every 10 requests, generate a new proxy
-    if n % 10 == 0:
-      proxy_index = random_proxy()
-      proxy = proxies[proxy_index]
-
-    # Make the call
-    try:
-      my_ip = urlopen(req).read().decode('utf8')
-      print('#' + str(n) + ': ' + my_ip)
-    except: # If error, delete this proxy and find another one
-      del proxies[proxy_index]
-      print('Proxy ' + proxy['ip'] + ':' + proxy['port'] + ' deleted.')
-      proxy_index = random_proxy()
-      proxy = proxies[proxy_index]
-
-# Retrieve a random index proxy (we need the index to delete it if not working)
 def random_proxy():
   return random.randint(0, len(proxies) - 1)
-
-if __name__ == '__main__':
-  main()
