+++ ds_scraper/scraper_to_get_articles_from_ds.py
@@ -1,150 +1,10 @@
-import os
 import time as pytime
-import re
-from urllib.request import urlopen
+import requests
 from bs4 import BeautifulSoup
 from datetime import date, timedelta
 
-base = "https://www.thedailystar.net/"
-numOfDaysToScrape = 8
-
-
-class Section:
-    """ 
-    Represents a section for Daily Star i.e. Front Page, Sports etc
-    """
-    def __init__(self, section_title,articles_raw,date):
-        """ 
-        Constructor
-        
-        @param {string} section_title - Title of the section
-        @param {[BeautifulSoup obj]} articles_raw - A list of BeautifulSoup obj. Each obj contains all the articles for a section
-        @param {string} date - Date of the articles
-        """
-        self.section_title = section_title
-        self.articles_raw = articles_raw
-        self.date = date
-        
-    def writeArticlesToFile(self):
-        """ 
-        Extracts and transforms each article from BeautifulSoup to an Article obj, then writes the article text to file
-        """
-        counter = 0
-        for article in self.articles_raw:
-            try:
-                article_title = article.h5.text
-                article_url = base + article.a['href']
-                article_html = urlopen(article_url)
-                article_soup = BeautifulSoup(article_html, 'html.parser')
-                articleObj = Article(self.section_title, article_title,article_soup, self.date)
-                articleObj.writeToFile()
-                counter += 1
-            except Exception as e:             
-                print(f"Error writing {article_title}")
-                print(f"Error message: {e}")
-                continue
-        
-        waitTime = 70
-        
-        print(f"Wrote {counter} articles to file")
-        print(f"Waiting {waitTime} seconds before making new requests...")
-        pytime.sleep(waitTime)
-        
-            
-
-class Article:
-    """
-    Represents a Daily Star article
-    """
-    def __init__(self,section_title,article_title,article_content, date):
-        """ 
-        Constructor
-        
-         @param {string} section_title - Title of the section
-         @param {string} article_title - Title of the article
-         @param {BeautifulSoup obj} article_content - A BeautifulSoup obj that encapsulates the contents of the article
-         @param {string} date - Date of the articles
-        """
-        self.section_title = section_title
-        self.article_title = article_title
-        self.article_content = article_content
-        self.date = date
-    
-    def getSectionTitle(self):
-        """ 
-        Returns the section_title
-        """
-        return self.section_title
-    
-    def getArticleTitle(self):
-        """ 
-        Returns the aritlce_title
-        """
-        return self.article_title
-    
-    def getArticleContent(self):
-        """ 
-        Returns the article_content
-        """
-        return self.article_content
-    
-    def extractText(self):
-        """
-        Extracts the articles from article_content and transforms it from BeautifulSoup obj to string
-        
-        @returns Article text in string format
-        """
-        paras = self.article_content.find('div',attrs={'class':["field-body view-mode-full","field-body"]}).find_all('p')
-        text = self._extractParasAndFormText(paras)
-        return text
-    
-    def _extractParasAndFormText(self,paras):
-        """
-        Extracts the string text from BeautifulSoup paragraphs
-        
-        @param {BeautifulSoup obj} paras - Contains article text
-        @returns Article text in string format
-        """
-        
-        text = "\n ".join([str(para.text) for para in paras])
-        return text
-    
-    def _createDirectories(self,dirPath):
-        """ 
-        Creates directories if they do not exist
-        
-        @param {string} dirPath - Directory path
-        """
-        if not os.path.isdir(dirPath):
-            os.makedirs(dirPath)
-            
-    def _sanitizeText(self,text):
-        """
-        Removes undesireable characters from the text so that they are valid file name.
-        
-        @param {string} text - tex tto be sanitized
-        """
-        text = text.strip().lower()
-        textWithoutSpecialChar = re.sub("[.?!\"\n',:]","",text)
-        return re.sub("[ ]","_",textWithoutSpecialChar)
-            
-    def writeToFile(self):
-        """ 
-        Write the article to a txt file.
-        """
-        dirPath  = self._sanitizeText(f"{self.date}/{self.section_title}")
-        self._createDirectories(dirPath)
-    
-        fileName = self._sanitizeText(f'{self.date}/{self.section_title}/{self.article_title}') + ".txt"
-    
-        with open(fileName,"w+", encoding="utf-8") as fobj:
-            #print(f"{self.section_title}: {self.article_title}")
-            fobj.write(self.extractText())
-        
-        
-        
-        
-### HELPER METHODS        
+from Section import Section
+from get_fake_user_proxies import getProxy       
 
 def findSectionsAndArticlesFromContent(date, content):
     """ 
@@ -173,7 +33,7 @@ def populateSections(date, articlesInSection):
     return sections
 
 
-def getContentAndTimeList(numOfDaysToScrape):
+def getContentAndTimeList():
     """Retrieves a list of content from Daily Star and a list of dates for which to scrape""" 
     content_list = []
     
@@ -190,8 +50,8 @@ def getDatesToScrape():
     Start and end dates need to changed manually inside the method.
     """
 
-    end_date = date(2018, 7, 21)  
-    star_date = date(2018, 7, 21)  
+    end_date = date(2018, 7, 20)  
+    star_date = date(2018, 7, 12)  
     dates = [star_date + timedelta(days=x) for x in range((end_date-star_date).days + 1)]
 
     return [str(d) for d in dates]
@@ -205,11 +65,23 @@ def getContent(date):
     """
     address = "https://www.thedailystar.net/newspaper?date={}".format(date)
     
-    html = urlopen(address)
-    soup = BeautifulSoup(html, 'html.parser')
+    response = makeSpoofedRequest(address)
+    #spoofedIp = getProxy()['ip']
+    #proxy = {"http": str(spoofedIp)}
+    #response = requests.get(address, proxies=proxy)
+    soup = BeautifulSoup(response.text, 'html.parser')
+    
     content = soup.find_all('div',attrs={'class':"panel-pane pane-news-col no-title block"}) 
 
     return content
+
+def makeSpoofedRequest(address):
+    spoofedIp = getProxy()['ip']
+    proxy = {"http": str(spoofedIp)}
+    response = requests.get(address, proxies=proxy)
+
+    return response        
+    
 def keepTrackOfScrapedArticles(date):
     """
     Writes the date to file a file. The date represents successfull scraping for that date
@@ -220,7 +92,7 @@ def keepTrackOfScrapedArticles(date):
             
 ### MAIN        
 def main():        
-    content_list, time_list = getContentAndTimeList(numOfDaysToScrape)
+    content_list, time_list = getContentAndTimeList()
         
     for i in range(len(content_list)):
         start = pytime.time()
